[BF] Delete toTransform invocation by reflection, disable cache due to some bugs
