+++ plugins/drouter-plugin/src/main/kotlin/com/didi/drouter/DRouterPlugin.kt
@@ -4,10 +4,8 @@ import com.android.build.api.artifact.ScopedArtifact
 import com.android.build.api.variant.AndroidComponentsExtension
 import com.android.build.api.variant.ScopedArtifacts
 import com.didi.drouter.plugin.RouterSetting
-import com.didi.drouter.utils.Logger
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.tasks.TaskProvider
 import org.gradle.configurationcache.extensions.capitalized
 import java.io.File
 
@@ -21,13 +19,14 @@ class DRouterPlugin : Plugin<Project> {
 
         val cacheDir = File(project.layout.buildDirectory.get().asFile, "intermediates/drouter")
         val isWindow = System.getProperty("os.name").startsWith("Windows")
-        val dRouterSettings: RouterSetting = project.extensions.create("drouter", RouterSetting::class.java)
+        val dRouterSettings: RouterSetting =
+            project.extensions.create("drouter", RouterSetting::class.java)
 
         val androidComponents = project.extensions.getByType(AndroidComponentsExtension::class.java)
         androidComponents.onVariants { variant ->
 
             val taskName = "${variant.name.capitalized()}DRouterTask"
-            val taskContainer: TaskProvider<RouterTransform> = project.tasks.register(
+            val taskContainer = project.tasks.register(
                 taskName,
                 RouterTransform::class.java,
                 androidComponents.sdkComponents.bootClasspath,
@@ -36,48 +35,13 @@ class DRouterPlugin : Plugin<Project> {
                 isWindow,
                 dRouterSettings,
             )
-
-            try {
-
-                val scopedArtifactsOperationClass = Class.forName(
-                    "com.android.build.api.artifact.impl.ScopedArtifactsImpl\$ScopedArtifactsOperationImpl"
-                )
-                val toTransformMethod = scopedArtifactsOperationClass.declaredMethods.find {
-                    it.name.contains("toTransform") && it.parameterCount == 5
-                }
-                toTransformMethod?.isAccessible = true
-                val use = variant.artifacts.forScope(ScopedArtifacts.Scope.ALL)
-                    .use(taskContainer)
-                toTransformMethod?.invoke(
-                    use,
+            variant.artifacts.forScope(ScopedArtifacts.Scope.ALL)
+                .use(taskContainer).toTransform(
                     ScopedArtifact.CLASSES,
                     RouterTransform::allJars,
                     RouterTransform::allDirectories,
-                    RouterTransform::outputJars,
-                    RouterTransform::outputClasses,
+                    RouterTransform::output
                 )
-            } catch (e: Exception) {
-                Logger.e("DRouterPlugin reflection error, try use former jar task")
-                taskContainer.get().enabled = false
-
-                val taskContainerOld = project.tasks.register(
-                    "${taskName}Jar",
-                    RouterTransformJar::class.java,
-                    androidComponents.sdkComponents.bootClasspath,
-                    androidComponents.pluginVersion,
-                    cacheDir,
-                    isWindow,
-                    dRouterSettings,
-                )
-                variant.artifacts.forScope(ScopedArtifacts.Scope.ALL)
-                    .use(taskContainerOld).toTransform(
-                        ScopedArtifact.CLASSES,
-                        RouterTransformJar::allJars,
-                        RouterTransformJar::allDirectories,
-                        RouterTransformJar::output
-                    )
-            }
         }
-
     }
 }
