+++ plugins/drouter-plugin/src/main/kotlin/com/didi/drouter/RouterTransform.kt
@@ -5,17 +5,21 @@ import com.didi.drouter.plugin.RouterSetting
 import com.didi.drouter.plugin.RouterTask
 import com.didi.drouter.utils.Logger
 import org.gradle.api.DefaultTask
-import org.gradle.api.file.ConfigurableFileCollection
-import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.file.Directory
 import org.gradle.api.file.RegularFile
+import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.ListProperty
-import org.gradle.api.tasks.Classpath
-import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
+import java.io.BufferedOutputStream
 import java.io.File
-import java.nio.file.Files
+import java.io.FileOutputStream
 import java.util.Queue
 import java.util.concurrent.ConcurrentLinkedQueue
+import java.util.jar.JarEntry
+import java.util.jar.JarFile
+import java.util.jar.JarOutputStream
 import javax.inject.Inject
 
 /**
@@ -29,16 +33,14 @@ abstract class RouterTransform @Inject constructor(
     private val dRouterSettings: RouterSetting,
 ) : DefaultTask() {
 
-    @get:Classpath
-    abstract val allJars: ConfigurableFileCollection
+    @get:InputFiles
+    abstract val allJars: ListProperty<RegularFile>
 
-    @get:Classpath
-    abstract val allDirectories: ConfigurableFileCollection
+    @get:InputFiles
+    abstract val allDirectories: ListProperty<Directory>
 
-    @get:OutputDirectory
-    abstract val outputJars: DirectoryProperty
-    @get:OutputDirectory
-    abstract val outputClasses: DirectoryProperty
+    @get:OutputFile
+    abstract val output: RegularFileProperty
 
     private val inputFiles = mutableSetOf<File>()
     private val routerDir = File(cacheDir, "router")
@@ -47,19 +49,19 @@ abstract class RouterTransform @Inject constructor(
     fun taskAction() {
         val timeStart = System.currentTimeMillis()
         Logger.v("DRouterTask start | AndroidGradlePlugin version: ${pluginVersion.run { "$major.$minor.$micro" }}")
+
         // 入口
         RouterTask(
-            assembleAllFile(), mutableSetOf(), dRouterSettings.cache, routerDir,
-            isWindows, cacheDir,
-            dRouterSettings,
+            assembleAllFile(),
+            mutableSetOf(),
+            false,
+            routerDir,
+            isWindows,
+            cacheDir,
+            dRouterSettings
         ).run(androidBootClasspath)
-//        val time = System.currentTimeMillis()
-        // 测试耗时 31 ms
-        copyToOutput()
-//        Logger.v("copyToOutput time: ${System.currentTimeMillis() - time}ms")
-        // 测试耗时 3 ms 但编译报错
-//        linkToOutput()
-//        Logger.v("linkToOutput time: ${System.currentTimeMillis() - time}ms")
+        inputFiles.add(routerDir)
+        packOutputJar()
 
         Logger.v("Link: https://github.com/didi/DRouter")
         Logger.v("DRouterTask done, time used: " + (System.currentTimeMillis() - timeStart) / 1000f  + "s")
@@ -71,74 +73,46 @@ abstract class RouterTransform @Inject constructor(
         androidBootClasspath.get().forEach {
             allFileQueue.add(it.asFile)
         }
-        allJars.files.forEach {
-            allFileQueue.add(it)
-            inputFiles.add(it)
+        allJars.get().forEach {
+            allFileQueue.add(it.asFile)
+            inputFiles.add(it.asFile)
         }
-        allDirectories.files.forEach {
-            allFileQueue.add(it)
-            inputFiles.add(it)
+        allDirectories.get().forEach {
+            allFileQueue.add(it.asFile)
+            inputFiles.add(it.asFile)
         }
         return allFileQueue
     }
 
-    private fun copyToOutput() {
-        val jarDirFile = outputJars.get().asFile
-        if(jarDirFile.exists()) {
-            jarDirFile.deleteRecursively()
-        }
-        jarDirFile.mkdirs()
-        allJars.files.forEachIndexed {index, it->
-            try {
-                it.copyTo(File(jarDirFile, it.name), false)
-            } catch (e: FileAlreadyExistsException) {
-                it.copyTo(File(jarDirFile, "${it.nameWithoutExtension}$index.${it.extension}"), false)
+    private fun packOutputJar() {
+        val jarOutput = JarOutputStream(BufferedOutputStream(FileOutputStream(output.get().asFile)))
+        val insertTag = mutableSetOf<String>()
+        inputFiles.forEach { input ->
+            if (input.isFile && input.name.lowercase().endsWith(".jar")) {
+                val jarFile = JarFile(input)
+                for (jarEntry in jarFile.entries()) {
+                    if (!insertTag.contains(jarEntry.name)) {
+                        insertTag.add(jarEntry.name)
+                        jarOutput.putNextEntry(JarEntry(jarEntry.name))
+                        jarFile.getInputStream(jarEntry).use {
+                            it.copyTo(jarOutput)
+                        }
+                        jarOutput.closeEntry()
+                    }
+                }
+                jarFile.close()
+            } else {
+                // input is dir
+                input.walk().filter { it.isFile }.forEach { file ->
+                    val relativePath = input.toURI().relativize(file.toURI()).path
+                    jarOutput.putNextEntry(JarEntry(relativePath.replace(File.separatorChar, '/')))
+                    file.inputStream().use { inputStream ->
+                        inputStream.copyTo(jarOutput)
+                    }
+                    jarOutput.closeEntry()
+                }
             }
         }
-
-        val classDirFile = outputClasses.get().asFile
-        if(classDirFile.exists()) {
-            classDirFile.deleteRecursively()
-        }
-        classDirFile.mkdirs()
-
-        allDirectories.forEach { dir->
-            dir.copyRecursively(classDirFile, false)
-        }
-
-        routerDir.copyRecursively(classDirFile, false)
+        jarOutput.close()
     }
-
-    private fun linkToOutput() {
-        val jarDirFile = outputJars.get().asFile
-        if(jarDirFile.exists()) {
-            jarDirFile.deleteRecursively()
-        }
-        jarDirFile.mkdirs()
-        allJars.files.forEach {
-            Files.createSymbolicLink(
-                File(jarDirFile, it.name).toPath(),
-                it.toPath(),
-            )
-        }
-
-        val classDirFile = outputClasses.get().asFile
-        if(classDirFile.exists()) {
-            classDirFile.deleteRecursively()
-        }
-        classDirFile.mkdirs()
-
-        allDirectories.forEach { dir->
-            Files.createSymbolicLink(
-                File(classDirFile, dir.name).toPath(),
-                dir.toPath(),
-            )
-        }
-
-        Files.createSymbolicLink(
-            File(classDirFile, routerDir.name).toPath(),
-            routerDir.toPath(),
-        )
-    }
-
 }
